backend/requirements.txt
Django==4.2.27
djangorestframework==3.14.0
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0
drf-nested-routers==0.93.4
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
django-extensions==3.2.3
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
setuptools>=65.5.1
reportlab==3.6.13
xhtml2pdf==0.2.11
openai>=1.0.0
